[ttf] split TTF and TTFhandle classes
TTF.h has special forward declaration for Freetype types,
which not really match to real implementation.

Therefore move TTFHandle from TTF.h and fully avoid
redefinition of Freetype types in header file.

Thus one can correctly use Freetype types and methods

Require immediate changes in TMathText and TPadPainterBase classes.
