+++ graf2d/graf/inc/TTF.h
@@ -1,7 +1,7 @@
 // @(#)root/graf:$Id$
 // Author: Olivier Couet     01/10/2002
 // Author: Fons Rademakers   21/11/1998
-// Author: Sergey Linev      29/14/2026
+// Author: Sergey Linev      29/04/2026
 
 /*************************************************************************
  * Copyright (C) 1995-2026, Rene Brun and Fons Rademakers.               *
@@ -14,9 +14,8 @@
 #ifndef ROOT_TTF
 #define ROOT_TTF
 
-
 #include "Rtypes.h"
-#include <memory>
+
 
 /// @cond DOXYGEN_IGNORE
 // Forward declare for the headers:
@@ -49,8 +48,9 @@ extern "C" {
 /// @endcond
 
 
-class TTFhandle;
-struct TTFontHandle;
+#include "TTFhandle.h"
+
+
 
 class TTF {
 
@@ -68,7 +68,7 @@ public:
       FT_Vector  fPos;          ///< position of glyph origin
       FT_Glyph   fImage{nullptr}; ///< glyph image
       TTGlyph(UInt_t indx = 0) : fIndex(indx) {}
-      ~TTGlyph();
+      ~TTGlyph() {}
    };
 
    TTF() {}
@@ -115,77 +115,6 @@ public:
    static void           CleanupGlyphs();
 
    ClassDef(TTF,0)  //Interface to TTF font handling
-};
-
-class TTFhandle {
-   friend class TTF;
-
-   private:
-      TTFontHandle  *fFont = nullptr;            ///< selected font
-      Int_t          fAscent = 0;                ///< string ascent, used to compute Y alignment
-      FT_BBox        fCBox;                      ///< string control box
-      std::vector<TTF::TTGlyph> fGlyphs;         ///< glyphs
-      Bool_t         fKerning = kTRUE;           ///< use kerning (true by default)
-      std::unique_ptr<FT_Matrix> fRotMatrix;     ///< rotation matrix
-      Int_t          fTBlankW = 0;               ///< trailing blanks width
-      Int_t          fWidth = 0;                 ///< string width, used to compute X alignment
-
-      static  Bool_t fgHinting;                   ///< use hinting (false by default)
-      static  Bool_t fgSmoothing;                 ///< use anti-aliasing (true when >8 planes, false otherwise)
-
-      UInt_t         CharToUnicode(UInt_t code);
-      void           ComputeTrailingBlanksWidth(Int_t n);
-
-      Int_t          SelectFontHandle(Int_t arg, const char *name = nullptr);
-
-      /// Thread-local wrapper to the FreeType library
-      struct FT_Library_Wrapper;
-      static thread_local FT_Library_Wrapper fFT_Library;
-
-   public:
-      TTFhandle();
-      virtual ~TTFhandle();
-
-      TTF::TTGlyph  *GetGlyphs() { return fGlyphs.data(); }
-      UInt_t         GetNumGlyphs() const { return fGlyphs.size(); }
-      FT_BitmapGlyph GetGlyphBitmap(UInt_t n, Bool_t smooth = kFALSE);
-      FT_Face        GetFontFace() const;
-      Int_t          GetAscent() const { return fAscent; }
-      Bool_t         GetKerning() const { return fKerning; }
-      FT_Matrix     *GetRotMatrix() const { return fRotMatrix.get(); }
-      Int_t          GetTrailingBlanksWidth() const { return fTBlankW; }
-      Int_t          GetWidth() const { return fWidth; }
-      const FT_BBox &GetBox() const { return fCBox; }
-
-      void           SetKerning(Bool_t state) { fKerning = state; }
-      void           SetTextFont(Font_t fontnumber);
-      Int_t          SetTextFont(const char *fontname, Int_t italic = 0);
-      Bool_t         SetTextSize(Float_t textsize);
-
-      void           LayoutGlyphs();
-      void           PrepareString(const char *string);
-      void           PrepareString(const wchar_t *string);
-      void           SetRotationMatrix(Float_t angle);
-      Int_t          GetGlyphsWidth() const;
-      Int_t          GetGlyphsHeight() const;
-      Bool_t         ApplyAlignRotate(Int_t &px, Int_t &py, Int_t align, Int_t pad_width, Int_t pad_height);
-      Bool_t         GetGlyphData(UInt_t n, Int_t &offx, Int_t &offy, UChar_t *&buffer, UInt_t &width, UInt_t &rows, UInt_t &pitch);
-      void           CleanupGlyphs();
-
-      void           GetTextExtent(UInt_t &w, UInt_t &h, const char *text);
-      void           GetTextExtent(UInt_t &w, UInt_t &h, const wchar_t *text);
-      void           GetTextAdvance(UInt_t &a, const char *text);
-
-      void           Version(Int_t &major, Int_t &minor, Int_t &patch);
-
-      static Bool_t  Init();
-      static Bool_t  GetHinting();
-      static Bool_t  GetSmoothing();
-      static void    SetHinting(Bool_t state);
-      static void    SetSmoothing(Bool_t state);
-
-   ClassDef(TTFhandle, 0)  // Dynamic interface to TTF
-
 };
 
 #endif
